fix(ext): resolve static analysis issues across extensions
Fix PHPStan and Psalm errors introduced by batch implementations across
CMS, admin, analytics, and forum extensions.

- Add type guards for mixed values in ImportAnalyzer (PHPStan cast.string)
- Fix string interpolation with mixed types using concatenation
- Resolve admin controller return type mismatches
- Fix analytics middleware parameter type declarations
- Update forum repository method signatures for portability layer

+++ extensions/cms/src/Internal/Commerce/TaxCalculator.php
@@ -41,7 +41,7 @@ final readonly class TaxCalculator implements TaxCalculatorInterface
 
         foreach ($items as $item) {
             $rate = $reverseCharge ? 0.0 : $this->findTaxRate($item['taxCategory'], $billingCountry);
-            $taxableAmount = $item['amount'] * $item['quantity'];
+            $taxableAmount = (float) ($item['amount'] * $item['quantity']);
             $taxAmount = intval(round($taxableAmount * $rate));
 
             $taxLineItems[] = new TaxLineItem(
